requirements.txt
ultralytics
matplotlib
opencv-python
deep-sort-realtime
pillow
argparse
